+++ api-sqlite/src/main/kotlin/de/md5lukas/waypoints/api/sqlite/WaypointHolderImpl.kt
@@ -21,7 +21,7 @@ import kotlinx.coroutines.withContext
 import org.bukkit.Location
 import org.bukkit.permissions.Permissible
 
-internal open class WaypointHolderImpl(
+internal abstract class WaypointHolderImpl(
     internal val dm: DatabaseManager,
     override val type: Type,
     private val owner: UUID?,
@@ -29,48 +29,60 @@ internal open class WaypointHolderImpl(
 
   override suspend fun getFolders(): List<Folder> =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.select(
-            "SELECT * FROM folders WHERE type = ? AND owner IS ?;", type.name, owner?.toString()) {
+            "SELECT * FROM folders WHERE type = ? AND (? IS NULL OR owner IS ?);",
+            type.name,
+            ownerString,
+            ownerString) {
               FolderImpl(dm, this)
             }
       }
 
   override suspend fun getWaypoints(): List<Waypoint> =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.select(
-            "SELECT * FROM waypoints WHERE type = ? AND owner IS ? AND folder IS NULL;",
+            "SELECT * FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?) AND folder IS NULL;",
             type.name,
-            owner?.toString()) {
+            ownerString,
+            ownerString) {
               WaypointImpl(dm, this)
             }
       }
 
   override suspend fun getAllWaypoints(): List<Waypoint> =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.select(
-            "SELECT * FROM waypoints WHERE type = ? AND owner IS ?;",
+            "SELECT * FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?);",
             type.name,
-            owner?.toString()) {
+            ownerString,
+            ownerString) {
               WaypointImpl(dm, this)
             }
       }
 
   override suspend fun getWaypointsAmount(): Int =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.selectFirst(
-            "SELECT COUNT(*) FROM waypoints WHERE type = ? AND owner IS ?;",
+            "SELECT COUNT(*) FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?);",
             type.name,
-            owner?.toString()) {
+            ownerString,
+            ownerString) {
               getInt(1)
             }!!
       }
 
   override suspend fun getFoldersAmount(): Int =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.selectFirst(
-            "SELECT COUNT(*) FROM folders WHERE type = ? AND owner IS ?;",
+            "SELECT COUNT(*) FROM folders WHERE type = ? AND (? IS NULL OR owner IS ?);",
             type.name,
-            owner?.toString()) {
+            ownerString,
+            ownerString) {
               getInt(1)
             }!!
       }
@@ -89,10 +101,15 @@ internal open class WaypointHolderImpl(
       }
 
   override suspend fun createWaypoint(name: String, location: Location): Waypoint {
-    return createWaypointTyped(name, location, type)
+    return createWaypointTyped(name, location, type, owner)
   }
 
-  internal suspend fun createWaypointTyped(name: String, location: Location, type: Type): Waypoint =
+  protected suspend fun createWaypointTyped(
+      name: String,
+      location: Location,
+      type: Type,
+      creator: UUID?,
+  ): Waypoint =
       withContext(dm.asyncDispatcher) {
         val id = UUID.randomUUID()
         dm.connection.update(
@@ -100,7 +117,7 @@ internal open class WaypointHolderImpl(
             id.toString(),
             OffsetDateTime.now().toString(),
             type.name,
-            owner?.toString(),
+            creator?.toString(),
             name,
             location.world!!.name,
             location.x,
@@ -114,7 +131,9 @@ internal open class WaypointHolderImpl(
             .also { WaypointCreateEvent(!dm.testing, it).callEvent() }
       }
 
-  override suspend fun createFolder(name: String): Folder =
+  override suspend fun createFolder(name: String): Folder = createFolder0(name, owner)
+
+  protected suspend fun createFolder0(name: String, creator: UUID?): Folder =
       withContext(dm.asyncDispatcher) {
         val id = UUID.randomUUID()
         dm.connection.update(
@@ -122,7 +141,7 @@ internal open class WaypointHolderImpl(
             id.toString(),
             OffsetDateTime.now().toString(),
             type.name,
-            owner?.toString(),
+            creator?.toString(),
             name,
         )
         dm.connection
@@ -136,10 +155,12 @@ internal open class WaypointHolderImpl(
 
   override suspend fun isDuplicateWaypointName(name: String): Boolean =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.selectFirst(
-            "SELECT EXISTS(SELECT 1 FROM waypoints WHERE type = ? AND owner IS ? AND name = ? COLLATE NOCASE);",
+            "SELECT EXISTS(SELECT 1 FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?) AND name = ? COLLATE NOCASE);",
             type.name,
-            owner?.toString(),
+            ownerString,
+            ownerString,
             name,
         ) {
           getInt(1) == 1
@@ -148,10 +169,12 @@ internal open class WaypointHolderImpl(
 
   override suspend fun isDuplicateFolderName(name: String): Boolean =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
         dm.connection.selectFirst(
-            "SELECT EXISTS(SELECT 1 FROM folders WHERE type = ? AND owner IS ? AND name = ? COLLATE NOCASE);",
+            "SELECT EXISTS(SELECT 1 FROM folders WHERE type = ? AND (? IS NULL OR owner IS ?) AND name = ? COLLATE NOCASE);",
             type.name,
-            owner?.toString(),
+            ownerString,
+            ownerString,
             name,
         ) {
           getInt(1) == 1
@@ -163,13 +186,16 @@ internal open class WaypointHolderImpl(
       permissible: Permissible?
   ): List<SearchResult<out Folder>> =
       withContext(dm.asyncDispatcher) {
+        val ownerString = owner?.toString()
+
         val (reducedQuery, taggedIndex) = prepareQuery(query)
 
         dm.connection
             .selectNotNull<Folder>(
-                "SELECT * FROM folders WHERE type = ? AND owner = ? AND name LIKE ? ESCAPE '!';",
+                "SELECT * FROM folders WHERE type = ? AND (? IS NULL OR owner IS ?) AND name LIKE ? ESCAPE '!';",
                 type.name,
-                owner?.toString(),
+                ownerString,
+                ownerString,
                 "$reducedQuery%",
             ) {
               FolderImpl(dm, this).also {
@@ -198,11 +224,13 @@ internal open class WaypointHolderImpl(
               val waypoint = "${result[1]}%"
 
               dm.connection.selectNotNull<Waypoint>(
-                  "SELECT * FROM waypoints WHERE type = ? AND owner IS ? AND name LIKE ? ESCAPE '!' AND folder IN (SELECT id FROM folders WHERE type = ? AND owner IS ? AND name LIKE ? ESCAPE '!');",
+                  "SELECT * FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?) AND name LIKE ? ESCAPE '!' AND folder IN (SELECT id FROM folders WHERE type = ? AND (? IS NULL OR owner IS ?) AND name LIKE ? ESCAPE '!');",
                   typeString,
+                  ownerString,
                   ownerString,
                   waypoint,
                   typeString,
+                  ownerString,
                   ownerString,
                   folder,
               ) {
@@ -215,8 +243,9 @@ internal open class WaypointHolderImpl(
               }
             } else {
               dm.connection.selectNotNull<Waypoint>(
-                  "SELECT * FROM waypoints WHERE type = ? AND owner IS ? AND name LIKE ? ESCAPE '!';",
+                  "SELECT * FROM waypoints WHERE type = ? AND (? IS NULL OR owner IS ?) AND name LIKE ? ESCAPE '!';",
                   typeString,
+                  ownerString,
                   ownerString,
                   "$reducedQuery%",
               ) {