+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/config/general/FeaturesConfiguration.kt
@@ -1,5 +1,6 @@
 package de.md5lukas.waypoints.config.general
 
+import de.md5lukas.konfig.ConfigPath
 import de.md5lukas.konfig.Configurable
 
 @Configurable
@@ -13,4 +14,13 @@ class FeaturesConfiguration {
 
   var teleportation = true
     private set
+
+  @ConfigPath("publicOwnership.waypoints")
+  var publicOwnershipWaypoints = false
+    private set
+
+  @ConfigPath("publicOwnership.folders")
+  var publicOwnershipFolders = false
+    private set
+    get() = field && publicOwnershipWaypoints
 }