+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/gui/pages/WaypointPage.kt
@@ -57,7 +57,10 @@ class WaypointPage(wpGUI: WaypointsGUI, private val waypoint: Waypoint) :
       when (waypoint.type) {
         Type.PRIVATE,
         Type.DEATH -> wpGUI.viewerData.id == waypoint.owner
-        Type.PUBLIC -> wpGUI.viewer.hasPermission(WaypointsPermissions.MODIFY_PUBLIC)
+        Type.PUBLIC ->
+            wpGUI.viewer.hasPermission(WaypointsPermissions.MODIFY_PUBLIC) ||
+                (wpGUI.plugin.waypointsConfig.general.features.publicOwnershipWaypoints &&
+                    wpGUI.viewerData.id == waypoint.owner)
         Type.PERMISSION -> wpGUI.viewer.hasPermission(WaypointsPermissions.MODIFY_PERMISSION)
       }